WIP Refactor `resolve` fn to be more flexible and extensible
Move node-specific resolution logic from `resolve` function to each node type class to allow extending / changing the behaviour in custom types, as well as cleaning up the code a bit
