+++ src/expression/node/ParenthesisNode.js
@@ -47,6 +47,16 @@ export const createParenthesisNode = /* #__PURE__ */ factory(name, dependencies,
     }
 
     /**
+     * TODO
+     * @param {*} scope 
+     * @param {*} within 
+     * @returns 
+     */
+    resolve (scope, within = new Set()) {
+      return new ParenthesisNode(this.content.resolve(scope, within))
+    }
+
+    /**
      * Get the content of the current Node.
      * @return {Node} content
      * @override
